:art: Separate authentication functions in their own useAuth hook
